+++ packages/button/src/types.ts
@@ -97,18 +97,21 @@ export interface ButtonProps {
    * The content that will appear inside of the `<Button />` component.
    */
   children?: React.ReactNode;
+
   /**
    * An icon glyph rendered before the button text.
    * To use a custom icon, see {@link Icon} {@link https://github.com/mongodb/leafygreen-ui/blob/main/packages/icon/README.md#usage-registering-custom-icon-sets | createIconComponent} docs
    * @type Leafygreen <Icon /> Component
    */
   leftGlyph?: React.ReactElement;
+
   /**
    * An icon glyph rendered after the button text.
    * To use a custom icon, see {@link Icon} {@link https://github.com/mongodb/leafygreen-ui/blob/main/packages/icon/README.md#usage-registering-custom-icon-sets | createIconComponent} docs
    * @type Leafygreen <Icon /> Component
    */
   rightGlyph?: React.ReactElement;
+
   /**
    * A `href` prop that will make the Button render as an anchor tag.
    */