+++ packages/loading-indicator/src/index.ts
@@ -1,0 +1,3 @@
+import PageLoader, { type PageLoaderProps } from './PageLoader';
+import Spinner, { type SpinnerProps } from './Spinner';
+export { PageLoader, type PageLoaderProps, Spinner, type SpinnerProps };