+++ packages/segmented-control/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @leafygreen-ui/segmented-control
 
+## 8.1.2
+
+### Patch Changes
+
+- 63b2deb0b: Upgrade button to v20.1.1
+
+## 8.1.1
+
+### Patch Changes
+
+- d2ce54e2f: Updates story files for Storybook 7.x
+- Updated dependencies [d2ce54e2f]
+- Updated dependencies [75099c60b]
+- Updated dependencies [d2ce54e2f]
+- Updated dependencies [0cd471676]
+  - @leafygreen-ui/box@3.1.4
+  - @leafygreen-ui/hooks@7.7.4
+  - @leafygreen-ui/icon@11.16.1
+  - @leafygreen-ui/leafygreen-provider@3.1.3
+  - @leafygreen-ui/lib@10.3.4
+  - @leafygreen-ui/typography@16.5.0
+
 ## 8.1.0
 
 ### Minor Changes